requirements.txt
Flask
Flask-WTF
Flask-Bootstrap
nltk
spacy
scikit-learn
pandas
PyPDF2
python-docx
transformers
torch